Intorduce PagedFluxBase
